Refactoring of the L2 switch application (see #5634)

+++ cherryd/openflow/of13/port.go
@@ -29,8 +29,8 @@ type Port struct {
 	currentSpeed, maxSpeed               uint32
 }
 
-func (r Port) Number() uint {
-	return uint(r.number)
+func (r Port) Number() uint32 {
+	return r.number
 }
 
 func (r Port) MAC() net.HardwareAddr {
